Uses an input domain text file

+++ DistanceMetric.java
@@ -1,6 +1,6 @@
 public class DistanceMetric {
     
-    public static float euclideanDistance(int valueOne, int valueTwo){
+    public static float euclideanDistance(float valueOne, float valueTwo){
         return Math.abs(valueOne - valueTwo);
     }
 }
